Add more output for install command
